+++ env/run.js
@@ -7,12 +7,15 @@ const express = require("express");
 const ws = require("ws");
 const http = require("http");
 const Database = require("@replit/database");
+const multiplayer = require("./multiplayer");
 const db = new Database();
 const app = express();
 const server = http.createServer(app);
-const wsServer = new ws.Server({ server: server, path: "/devws" });
 const port = process.env.PORT || 8000;
 let err = null;
+
+// start multiplayer server
+multiplayer(server);
 
 // build user game
 function buildGame() {
@@ -41,7 +44,7 @@ function buildGame() {
 			sourcemap: true,
 			target: "es6",
 			keepNames: true,
-			entryPoints: ["helper.js"],
+			entryPoints: ["helper.ts"],
 			outfile: "dist/helper.js",
 		});
 
@@ -100,10 +103,18 @@ app.get("/user", (req, res) => {
 	}
 });
 
-// TODO: authed user level abstraction?
 app.get("/db", async (req, res) => {
 	try {
 		res.json(await db.list());
+	} catch (e) {
+		res.sendStatus(500);
+	}
+});
+
+app.delete("/db", async (req, res) => {
+	try {
+		await db.empty();
+		res.sendStatus(200);
 	} catch (e) {
 		res.sendStatus(500);
 	}
@@ -119,7 +130,16 @@ app.get("/db/:item", async (req, res) => {
 
 app.post("/db/:item", async (req, res) => {
 	try {
-		db.set(req.params.item, req.body);
+		await db.set(req.params.item, req.body);
+		res.sendStatus(200);
+	} catch (e) {
+		res.sendStatus(500);
+	}
+});
+
+app.delete("/db/:item", async (req, res) => {
+	try {
+		await db.delete(req.params.item);
 		res.sendStatus(200);
 	} catch (e) {
 		res.sendStatus(500);