+++ site/land.js
@@ -8,17 +8,17 @@ const css = {
 		"font-family": "Necto Mono",
 	},
 	"body": {
-		"background": `url("/site/img/bg.svg") repeat`,
 		"padding": "48px",
+		"background": "var(--color-bg) url(/site/img/bg.svg) repeat",
 		"animation": "scroll linear 5s infinite",
 	},
 	"a": {
-		"color": "#333333",
+		"color": "var(--color-fg)",
 		"border-radius": "12px",
 		"padding": "2px 4px",
 		":hover": {
-			"background": "#333333",
-			"color": "white",
+			"background": "var(--color-fg)",
+			"color": "var(--color-bg)",
 		},
 	},
 	"#fun": {
@@ -91,6 +91,11 @@ const css = {
 			},
 		},
 	},
+	"label[for='themeswitch']": {
+		"position": "fixed",
+		"right": "16px",
+		"bottom": "16px",
+	},
 	"@keyframes": {
 		"scroll": {
 			"0%": {
@@ -157,7 +162,7 @@ const page = t("html", {}, [
 					", ",
 					t("a", { href: "/demo", }, "demos"),
 					", \n",
-					t("a", { href: "/tut/intro.md", }, "tutorial"),
+					t("a", { href: "/doc/intro.md", }, "tutorial"),
 					", and ",
 					t("a", { href: "/github", }, "github"),
 					".",
@@ -165,6 +170,12 @@ const page = t("html", {}, [
 				t("img", { id: "highlight", src: "/site/img/highlight.png", }),
 				t("img", { id: "outline", src: "/site/img/outline.png", }),
 				t("img", { id: "circle", src: "/site/img/circle.png", }),
+			]),
+		]),
+		t("input", { id: "themeswitch", type: "checkbox", name: "themeswitch", style: "display: none" }, ""),
+		t("label", { for: "themeswitch", class: "switch theme", }, [
+			t("div", { class: "strip", }, [
+				t("div", { class: "ball", }, []),
 			]),
 		]),
 		t("script", { src: "/site/js/land.js" }, ""),