+++ src/kaboom.ts
@@ -78,7 +78,7 @@ const app = appInit({
 });
 
 const gfx = gfxInit(app.gl, {
-	clearColor: gconf.clearColor ? rgb(gconf.clearColor) : undefined,
+	background: gconf.background ? rgb(gconf.background) : undefined,
 	width: gconf.width,
 	height: gconf.height,
 	scale: gconf.scale,
@@ -560,7 +560,6 @@ function action(tag: Tag | (() => void), cb?: (obj: Character) => void): EventCa
 	if (typeof tag === "function" && cb === undefined) {
 		return add([{ update: tag, }]).destroy;
 	} else if (typeof tag === "string") {
-		// TODO add an empty game obj
 		return on("update", tag, cb);
 	}
 }
@@ -570,7 +569,7 @@ function render(tag: Tag | (() => void), cb?: (obj: Character) => void) {
 	if (typeof tag === "function" && cb === undefined) {
 		return add([{ draw: tag, }]).destroy;
 	} else if (typeof tag === "string") {
-		return on("update", tag, cb);
+		return on("draw", tag, cb);
 	}
 }
 
@@ -1021,7 +1020,7 @@ function pos(...args): PosComp {
 			let dy = p.y;
 			let col = null;
 
-			if (this.solid) {
+			if (this.solid && this.area) {
 
 				let a1 = this.worldArea();
 
@@ -1286,68 +1285,36 @@ function follow(obj: Character, offset?: Vec2): FollowComp {
 }
 
 function move(direction: number | Vec2, speed: number): MoveComp {
-
 	const d = typeof direction === "number" ? dir(direction) : direction.unit();
-	let timeOut = 0;
-	// if it's not seen in 6 seconds, we destroy it
-	const maxTimeOut = 6;
-
-	function isOut(p: Vec2) {
-		let is = false;
-		if (d.x < 0) {
-			is ||= p.x < 0;
-		} else if (d.x > 0) {
-			is ||= p.x > width();
-		}
-		if (d.y < 0) {
-			is ||= p.y < 0;
-		} else if (d.y > 0) {
-			is ||= p.y > width();
-		}
-		return is;
-	}
-
 	return {
-
 		id: "move",
 		require: [ "pos", ],
-
 		update() {
-
-			// move
 			this.move(d.scale(speed));
-
-			// check if out of screen
-			const pos = this.screenPos();
-
-			if (isOut(pos)) {
-				if (this.width && this.height) {
-					const w = this.width;
-					const h = this.height;
-					const s = this.scale ?? vec2(1);
-					const orig = originPt(this.origin || DEF_ORIGIN);
-					const p1 = pos.sub(orig.sub(-1, -1).scale(0.5).scale(w, h).scale(s));
-					const p2 = pos.sub(orig.sub(1, 1).scale(0.5).scale(w, h).scale(s));
-					if (isOut(p1) && isOut(p2)) {
-						timeOut += dt();
-					} else {
-						timeOut = 0;
-					}
-				} else {
-					timeOut += dt();
-				}
-			} else {
-				timeOut = 0;
-			}
-
-			if (timeOut >= maxTimeOut) {
-				destroy(this);
-			}
-
 		},
-
 	};
+}
 
+function cleanup(time: number = 0): CleanupComp {
+	let timer = 0;
+	return {
+		id: "cleanup",
+		require: [ "pos", "area", ],
+		update() {
+			const screenRect = {
+				p1: vec2(0, 0),
+				p2: vec2(width(), height()),
+			}
+			if (colRectRect(this.screenArea(), screenRect)) {
+				timer = 0;
+			} else {
+				timer += dt();
+				if (timer >= time) {
+					this.destroy();
+				}
+			}
+		},
+	};
 }
 
 // TODO: tell which side collides
@@ -1682,8 +1649,8 @@ function sprite(id: string | SpriteData, conf: SpriteCompConf = {}): SpriteComp 
 							this.frame = anim.from;
 						} else {
 							this.frame++;
-							this.stop();
 							curAnim.onEnd();
+							this.stop();
 						}
 					}
 				} else {
@@ -1693,8 +1660,8 @@ function sprite(id: string | SpriteData, conf: SpriteCompConf = {}): SpriteComp 
 							this.frame = anim.from;
 						} else {
 							this.frame--;
-							this.stop();
 							curAnim.onEnd();
+							this.stop();
 						}
 					}
 				}
@@ -1940,10 +1907,10 @@ function body(conf: BodyCompConf = {}): BodyComp {
 	return {
 
 		id: "body",
-		require: [ "area", "pos", ],
+		require: [ "pos", ],
 		jumpForce: conf.jumpForce ?? DEF_JUMP_FORCE,
 		weight: conf.weight ?? 1,
-		solid: true,
+		solid: conf.solid ?? true,
 
 		update() {
 
@@ -2103,19 +2070,22 @@ function health(hp: number): HealthComp {
 	};
 }
 
-function lifespan(time: number, cb?: () => void): LifespanComp {
+function lifespan(time: number, conf: LifespanCompConf = {}): LifespanComp {
 	if (time == null) {
 		throw new Error("lifespan() requires time");
 	}
 	let timer = 0;
+	const fade = conf.fade ?? 0;
+	const startFade = Math.max((time - fade), 0);
 	return {
 		id: "lifespan",
 		update() {
 			timer += dt();
+			// TODO: don't assume 1 as start opacity
+			if (timer >= startFade) {
+				this.opacity = map(timer, startFade, time, 1, 0);
+			}
 			if (timer >= time) {
-				if (cb) {
-					cb.call(this);
-				}
 				this.destroy();
 			}
 		},
@@ -2454,6 +2424,7 @@ const ctx: KaboomCtx = {
 	lifespan,
 	z,
 	move,
+	cleanup,
 	follow,
 	// group events
 	on,