+++ src/types.ts
@@ -14,7 +14,7 @@
  *     stretch: true,
  *     letterbox: true,
  *     font: "sinko",
- *     clearColor: [ 0, 0, 255, ],
+ *     background: [ 0, 0, 255, ],
  * });
  *
  * // all kaboom functions are imported to global automatically
@@ -412,6 +412,20 @@ interface KaboomCtx {
 	 */
 	move(direction: number | Vec2, speed: number): MoveComp,
 	/**
+	 * destroy() the character if it's out of screen. Optionally specify the amount of time it has to be off-screen before removal.
+	 *
+	 * @example
+	 * ```js
+	 * // remove this 3 seconds after it leaves screen
+	 * add([
+	 *     pos(80, 80),
+	 *     move(LEFT, 120),
+	 *     cleanup(3),
+	 * ]);
+	 * ```
+	 */
+	cleanup(time?: number): CleanupComp,
+	/**
 	 * Follow another game obj's position.
 	 */
 	follow(obj: Character | null, offset?: Vec2): FollowComp,
@@ -495,7 +509,7 @@ interface KaboomCtx {
 	 * ]);
 	 * ```
 	 */
-	lifespan(time: number, action?: () => void): LifespanComp,
+	lifespan(time: number, conf?: LifespanCompConf): LifespanComp,
 	/**
 	 * Register an event on all game objs with certain tag.
 	 *
@@ -1406,7 +1420,6 @@ type MergeObj<T> = Expand<UnionToIntersection<Defined<T>>>;
 type MergeComps<T> = Omit<MergeObj<T>, keyof Comp>;
 
 type CompList<T> = Array<T | Tag>;
-type DynCompList<T> = CompList<T> | ((...args: any[]) => CompList<T>);
 
 type Key =
 	| "f1" | "f2" | "f3" | "f4" | "f5" | "f6" | "f7" | "f8" | "f9" | "f10" | "f11" | "f12"
@@ -1470,7 +1483,7 @@ interface KaboomConf {
 	/**
 	 * Background color. E.g. [ 0, 0, 255 ] for solid blue background.
 	 */
-	clearColor?: number[],
+	background?: number[],
 	/**
 	 * The color to draw collider boxes etc.
 	 */
@@ -2095,7 +2108,6 @@ interface Comp {
 type CharacterID = number;
 
 interface PosComp extends Comp {
-//  	id: "pos",
 	/**
 	 * Object's current world position.
 	 */
@@ -2122,7 +2134,6 @@ interface PosComp extends Comp {
 }
 
 interface ScaleComp extends Comp {
-//  	id: "scale",
 	scale: Vec2,
 	scaleTo(s: number): void,
 	scaleTo(s: Vec2): void,
@@ -2130,7 +2141,6 @@ interface ScaleComp extends Comp {
 }
 
 interface RotateComp extends Comp {
-//  	id: "rotate",
 	/**
 	 * Angle in degrees.
 	 */
@@ -2138,17 +2148,14 @@ interface RotateComp extends Comp {
 }
 
 interface ColorComp extends Comp {
-//  	id: "color",
 	color: Color,
 }
 
 interface OpacityComp extends Comp {
-//  	id: "opacity",
 	opacity: number,
 }
 
 interface OriginComp extends Comp {
-//  	id: "origin",
 	/**
 	 * Origin point for render.
 	 */
@@ -2156,7 +2163,6 @@ interface OriginComp extends Comp {
 }
 
 interface LayerComp extends Comp {
-//  	id: "layer",
 	/**
 	 * Which layer this game obj belongs to.
 	 */
@@ -2164,7 +2170,6 @@ interface LayerComp extends Comp {
 }
 
 interface ZComp extends Comp {
-//  	id: "z",
 	/**
 	 * Defines the z-index of this game obj
 	 */
@@ -2172,7 +2177,6 @@ interface ZComp extends Comp {
 }
 
 interface FollowComp extends Comp {
-//  	id: "follow",
 	follow: {
 		obj: Character,
 		offset: Vec2,
@@ -2180,7 +2184,9 @@ interface FollowComp extends Comp {
 }
 
 interface MoveComp extends Comp {
-//  	id: "move",
+}
+
+interface CleanupComp extends Comp {
 }
 
 type RectSide =
@@ -2214,7 +2220,6 @@ interface AreaCompConf {
 }
 
 interface AreaComp extends Comp {
-//  	id: "area",
 	/**
 	 * Collider area info.
 	 */
@@ -2325,7 +2330,6 @@ interface SpriteCompConf {
 }
 
 interface SpriteComp extends Comp {
-//  	id: "sprite",
 	/**
 	 * Width for sprite.
 	 */
@@ -2373,7 +2377,6 @@ interface SpriteComp extends Comp {
 }
 
 interface TextComp extends Comp {
-//  	id: "text",
 	/**
 	 * The text to render.
 	 */
@@ -2412,7 +2415,6 @@ interface TextCompConf {
 }
 
 interface RectComp extends Comp {
-//  	id: "rect",
 	/**
 	 * Width of rect.
 	 */
@@ -2432,7 +2434,6 @@ type AreaType =
 	;
 
 interface OutlineComp extends Comp {
-//  	id: "outline",
 	lineWidth: number,
 	lineColor: Color,
 }
@@ -2494,13 +2495,11 @@ type UniformValue =
 type Uniform = Record<string, UniformValue>;
 
 interface ShaderComp extends Comp {
-//  	id: "shader",
 	uniform: Uniform,
 	shader: string,
 }
 
 interface BodyComp extends Comp {
-//  	id: "body",
 	/**
 	 * If should collide with other solid objects.
 	 */
@@ -2528,7 +2527,7 @@ interface BodyComp extends Comp {
 	/**
 	 * Upward thrust.
 	 */
-	jump(f?: number): void,
+	jump(force?: number): void,
 	/**
 	 * Performs double jump (the initial jump only happens if player is grounded).
 	 */
@@ -2548,6 +2547,10 @@ interface BodyCompConf {
 	 * Gravity multiplier.
 	 */
 	weight?: number,
+	/**
+	 * If should not move through other solid objects.
+	 */
+	solid?: boolean,
 	/**
 	 * Can you hang to a wall.
 	 */
@@ -2574,7 +2577,6 @@ interface Timer {
 }
 
 interface TimerComp extends Comp {
-//  	id: "timer",
 	/**
 	 * Run the callback after n seconds.
 	 */
@@ -2582,7 +2584,6 @@ interface TimerComp extends Comp {
 }
 
 interface SolidComp extends Comp {
-//  	id: "solid",
 	/**
 	 * If should stop other solid objects from moving through.
 	 */
@@ -2590,7 +2591,6 @@ interface SolidComp extends Comp {
 }
 
 interface FixedComp extends Comp {
-//  	id: "fixed",
 	/**
 	 * If the obj is unaffected by camera
 	 */
@@ -2598,7 +2598,6 @@ interface FixedComp extends Comp {
 }
 
 interface StayComp extends Comp {
-//  	id: "stay",
 	/**
 	 * If the obj should not be destroyed on scene switch.
 	 */
@@ -2606,7 +2605,6 @@ interface StayComp extends Comp {
 }
 
 interface HealthComp extends Comp {
-//  	id: "health",
 	/**
 	 * Decrease HP by n (defaults to 1).
 	 */
@@ -2626,7 +2624,13 @@ interface HealthComp extends Comp {
 }
 
 interface LifespanComp extends Comp {
-//  	id: "lifespan",
+}
+
+interface LifespanCompConf {
+	/**
+	 * Fade out duration (default 0 which is no fade out).
+	 */
+	fade?: number,
 }
 
 interface LevelConf {